+++ src/jobs/MediaSync.php
@@ -116,6 +116,8 @@ class MediaSync extends BaseJob
                 $availabilities->public->start,
                 $availabilities->public->end
             );
+						
+						$isNew = !$existingEntry;
 
             // Set default field Values
             $defaultFields = [];
@@ -123,7 +125,7 @@ class MediaSync extends BaseJob
             // Set field values based on API Column Fields on settings
             $apiColumnFields = SettingsHelper::get( 'apiColumnFields' );
 						
-						if($this->fieldsToSync === '*' || in_array('title', $this->fieldsToSync)) {
+						if($this->fieldsToSync === '*' || in_array('title', $this->fieldsToSync) || $isNew ){
 								$entry->title = $assetAttributes->title;
 						}
 
@@ -132,7 +134,7 @@ class MediaSync extends BaseJob
                 $apiField = $apiColumnField[ 0 ];
 								
 								// ensure the field to be updated from MM Settings is included in the fieldsToSync array
-								if($this->fieldsToSync !== '*' && !in_array($apiField, $this->fieldsToSync) ) {
+								if(!$isNew && ($this->fieldsToSync !== '*' && !in_array($apiField, $this->fieldsToSync)) ) {
 									continue;
 								}
 							
@@ -214,10 +216,11 @@ class MediaSync extends BaseJob
                             if( $tag ) {
                                 array_push( $siteTags, $tag->id );
                             }
+
                         }
-
+												$this->siteTags = $siteTags;
                         $defaultFields[ $siteTagFieldHandle ] = $siteTags;
-
+												
                     break;
                     case 'film_tags':
 
@@ -248,12 +251,11 @@ class MediaSync extends BaseJob
                             }
 
                             if( $film ) {
-                                if( $tag ) {
-                                    array_push( $filmTags, $film->id );
-                                }
+                                $filmTags[] = $film->id;
                             }
                         }
-
+												
+												$this->filmTags = $filmTags;
                         $defaultFields[ $filmTagFieldHandle ] = $filmTags;
 
                     break;
@@ -282,6 +284,7 @@ class MediaSync extends BaseJob
                             }
                         }
 
+												$this->topicTags = $topicTags;
                         $defaultFields[ $topicTagFieldHandle ] = $topicTags;
 
                     break;